refactor: extract task dispatch into daemon/dispatch.rs
Move orphan recovery, duplicate worker detection, pending task
spawning, orphan worktree cleanup, and discovered-coworker nudging
from mod.rs into a dedicated dispatch.rs module. Updates call sites
in events.rs, rpc.rs, and mod.rs.

mod.rs shrinks by ~660 lines with no behavioral changes.
