+++ src/daemon/events.rs
@@ -62,9 +62,9 @@ pub async fn evaluate_tick(
         }
         DaemonEvent::TaskDispatchTick => {
             let mut effects = Vec::new();
-            effects.extend(super::check_for_duplicate_task_workers(snap));
-            effects.extend(super::check_and_recover_orphans(snap, state));
-            effects.extend(super::spawn_for_pending_tasks(snap, state));
+            effects.extend(super::dispatch::check_for_duplicate_task_workers(snap));
+            effects.extend(super::dispatch::check_and_recover_orphans(snap, state));
+            effects.extend(super::dispatch::spawn_for_pending_tasks(snap, state));
             effects.extend(super::health::check_and_respawn_zombies(snap, state));
             effects.extend(super::health::check_and_fire_reminders(snap, state));
             effects
